Lock file improvements.
Information relating to the process ID and process start time are now added to the lock file. This seems like a more robust / cross platform approach than dealing with conversions of the date format from ps. It is also a more robust approach than relying upon  the creation time of the lock file.